Move IDB interfaces from node/schema to chainkit/schema
The IDBTool and IDBChainkit interfaces were relocated from node/schema/idb.go to chainkit/schema/schema.go for better organization and to centralize schema definitions used by chainkit. References in chainkit/chainkit.go were updated accordingly.
