+++ chainkit/schema/schema.go
@@ -1,6 +1,9 @@
 package schema
 
-import goarSchema "github.com/permadao/goar/schema"
+import (
+	vmmSchema "github.com/hymatrix/hymx/vmm/schema"
+	goarSchema "github.com/permadao/goar/schema"
+)
 
 type Config struct {
 	RedisUrl     string `json:"redisUrl" yaml:"redisUrl"`
@@ -13,4 +16,27 @@ type DownloadResult struct {
 	Nonce      int64
 	Assignment *goarSchema.BundleItem
 	Message    *goarSchema.BundleItem
+}
+
+type IDBTool interface {
+	// DB
+	GetMessage(msgid string) (msg *goarSchema.BundleItem, err error)
+	GetAssignByNonce(pid string, nonce int64) (assign *goarSchema.BundleItem, err error)
+	GetResult(msgid string) (result *vmmSchema.Result, err error)
+}
+
+type IDBChainkit interface {
+	AddPending(txid string) error
+	MoveToUploading() (int64, error)
+	EndUpload() error
+
+	IsUploadedBatch(txids []string) (map[string]bool, error)
+	IsUploaded(txid string) (bool, error)
+
+	GetUploading() ([]string, error)
+	SetBundledIn(bundledInID string) error
+	GetBundledIn() (string, error)
+
+	Cache(txid string, item goarSchema.BundleItem) error
+	GetCache(txid string) (*goarSchema.BundleItem, error)
 }
