+++ node/schema/idb.go
@@ -39,26 +39,3 @@ type IDBOutbox interface {
 	Checkpoint(pid string) (string, error)
 	Restore(data string) error
 }
-
-type IDBTool interface {
-	// DB
-	GetMessage(msgid string) (msg *goarSchema.BundleItem, err error)
-	GetAssignByNonce(pid string, nonce int64) (assign *goarSchema.BundleItem, err error)
-	GetResult(msgid string) (result *vmmSchema.Result, err error)
-}
-
-type IDBChainkit interface {
-	AddPending(txid string) error
-	MoveToUploading() (int64, error)
-	EndUpload() error
-
-	IsUploadedBatch(txids []string) (map[string]bool, error)
-	IsUploaded(txid string) (bool, error)
-
-	GetUploading() ([]string, error)
-	SetBundledIn(bundledInID string) error
-	GetBundledIn() (string, error)
-
-	Cache(txid string, item goarSchema.BundleItem) error
-	GetCache(txid string) (*goarSchema.BundleItem, error)
-}
